Mobile Distro Prototyping: FE - make Distro Spot title of Event clickable
