+++ react/src/utils/hooks.ts
@@ -1,6 +1,22 @@
 import { useState, useCallback } from "react";
 import { useNavigate, useParams } from "react-router-dom";
 
+
+export const useGetUrlForResourceHelpers = () => {
+  const baseId = useParams<{ baseId: string }>().baseId;
+  const getDistroSpotDetailUrlById = (distroSpotId: string) =>
+    `/bases/${baseId}/distributions/spots/${distroSpotId}`;
+
+  const getDistroEventDetailUrlById = (distroEventId: string) =>
+    `/bases/${baseId}/distributions/events/${distroEventId}`;
+
+  return {
+    getDistroSpotDetailUrlById,
+    getDistroEventDetailUrlById,
+  };
+};
+
+
 export const useToggle = (initialValue = false) => {
   const [value, setValue] = useState(initialValue);
   const toggle = useCallback(() => {
